rag/requirements.txt
fastapi
uvicorn
psycopg2-binary
numpy
sentence-transformers
pgvector
